requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
asgiref==3.6.0
    # via flask
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via flask
flask[async]==2.2.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-sqlalchemy
    #   strawberry-graphql
flask-cors==3.0.10
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via -r requirements.in
graphql-core==3.2.3
    # via strawberry-graphql
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
python-dateutil==2.8.2
    # via strawberry-graphql
python-dotenv==1.0.0
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   flask-cors
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.7
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
strawberry-graphql[flask]==0.165.0
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   sqlalchemy
    #   strawberry-graphql
werkzeug==2.2.3
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
